mehmeh^12

+++ firmware/micromouse/lib/bno_driver/BNO055.cpp
@@ -2,8 +2,7 @@
 #include "Arduino.h"
 #include "BNO055_registers.h"
 
-imu::imu(const int8_t SCL, const int8_t SDA, const i2c_port_t port,
-         const uint8_t address) :
+imu::imu(const int8_t SCL, const int8_t SDA, const i2c_port_t port, const uint8_t address) :
     SCL(SCL), SDA(SDA), address(address), i2c(port == 0 ? Wire : Wire1) {}
 
 
@@ -41,13 +40,11 @@ void imu::remap_axis(const uint8_t config) const {
 
 void imu::remap_axis(Axis_remap_config config) const {
     this->set_mode(operation_mode::CONFIG);
-    const auto config_axis = static_cast<uint8_t>(config.x_axis) |
-        static_cast<uint8_t>(config.y_axis) << 2 |
+    const auto config_axis = static_cast<uint8_t>(config.x_axis) | static_cast<uint8_t>(config.y_axis) << 2 |
         static_cast<uint8_t>(config.z_axis) << 4;
     write_register(imu_registers::axis_remap::REMAP_AXIS, config_axis);
 
-    const auto config_sign = static_cast<uint8_t>(config.x_sign) << 2 |
-        static_cast<uint8_t>(config.y_sign) << 1 |
+    const auto config_sign = static_cast<uint8_t>(config.x_sign) << 2 | static_cast<uint8_t>(config.y_sign) << 1 |
         static_cast<uint8_t>(config.z_sign);
     write_register(imu_registers::axis_remap::REMAP_SIGN, config_sign);
     this->set_mode(operation_mode::NDOF);
@@ -81,8 +78,7 @@ vec_3 imu::acceleration() const {
     uint8_t buf[6] = {};
     read_register(imu_registers::sensor_data::ACCEL_X_LSB, buf, 6);
     for (int i = 0; i < 3; ++i) {
-        v.vec[i] = static_cast<float>(buf[2 * i + 1] << 8 | buf[2 * i]) /
-            100.0f; // scale: 1 LSB = 1 mg = 0.01 m/s^2
+        v.vec[i] = static_cast<float>(static_cast<int16_t>(buf[2 * i + 1] << 8 | buf[2 * i])) / 100.0f; // scale: 1 LSB = 1 mg = 0.01 m/s^2
     }
     return v;
 }
@@ -92,8 +88,7 @@ vec_3 imu::gyro() const {
     uint8_t buf[6] = {};
     read_register(imu_registers::sensor_data::GYRO_X_LSB, buf, 6);
     for (int i = 0; i < 3; ++i)
-        v.vec[i] =
-            static_cast<float>(buf[2 * i + 1] << 8 | buf[2 * i]) / 900.0f;
+        v.vec[i] = static_cast<float>(static_cast<int16_t>(buf[2 * i + 1] << 8 | buf[2 * i])) / 900.0f;
     return v;
 }
 
@@ -102,7 +97,7 @@ vec_3 imu::mag() const {
     uint8_t buf[6] = {};
     read_register(imu_registers::sensor_data::MAG_X_LSB, buf, 6);
     for (int i = 0; i < 3; ++i)
-        v.vec[i] = static_cast<float>(buf[2 * i + 1] << 8 | buf[2 * i]) / 16.0f;
+        v.vec[i] = static_cast<float>(static_cast<int16_t>(buf[2 * i + 1] << 8 | buf[2 * i])) / 16.0f;
     return v;
 }
 
@@ -127,7 +122,7 @@ void imu::set_yaw_offset() {
 float imu::relative_heading() const {
     uint8_t buf[2] = {};
     read_register(imu_registers::sensor_data::EUL_X_LSB, buf, 2);
-    float angle = static_cast<float>(buf[1] << 8 | buf[0]) / 16.0f;
+    float angle = static_cast<float>(static_cast<int16_t>(buf[1] << 8 | buf[0])) / 16.0f;
     angle = fmod(angle + 180 - this->yaw_offset, 360.0f);
     if (angle < 0)
         angle += 360;
@@ -145,8 +140,7 @@ vec_4 imu::quaternion() const {
     uint8_t buf[8];
     read_register(sensor_data::QUAT_W_LSB, buf, 8);
     for (int i = 0; i < 4; ++i)
-        v.vec[i] =
-            static_cast<float>(buf[2 * i + 1] << 8 | buf[2 * i]) / 16384.0f;
+        v.vec[i] = static_cast<float>(buf[2 * i + 1] << 8 | buf[2 * i]) / 16384.0f;
     return v;
 }
 
@@ -158,8 +152,7 @@ void imu::write_register(const uint8_t reg, const uint8_t val) const {
     i2c.endTransmission();
 }
 
-uint8_t imu::read_register(const uint8_t reg, uint8_t* buffer,
-                           const uint8_t len) const {
+uint8_t imu::read_register(const uint8_t reg, uint8_t* buffer, const uint8_t len) const {
     i2c.beginTransmission(address);
     i2c.write(reg);
     i2c.endTransmission(false);
